requirements.txt
Flask
numpy
torch
torchvision
Pillow
open-clip-torch
pandas
tqdm
scikit-learn